+++ .gitignore
@@ -106,6 +106,5 @@ ENV/
 # Dask
 dask-worker-space/
 
-
-# Generated conda requirements file
-_req-*.txt
+# Hatch-vcs generated version file
+_version.py
